Move CountriesViewSet to v2 file
